+++ CHANGELOG.md
@@ -31,6 +31,10 @@ Sections
 ### Fixed
 - `AcessoryDriver.add_job` now correctly schedules coroutines wrapped in functools.partial.
 
+### Reverted
+- Char.client_update_value no longer ignores duplicate values. Reverts [#162](https://github.com/ikalchev/HAP-python/pull/162). [#166](https://github.com/ikalchev/HAP-python/pull/166)
+
+
 ## [2.3.0] - 2018-10-25
 
 ### Added