+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Sections
 `reconfigure_stream` are now async.
 - The streaming process is started with `asyncio.create_subprocess_exec` instead of
 `subprocess.Popen`
+- Temporary disabled `update_advertisement` call after `pair` and `unpair`. To pair the accessory again after a successful unpair, you need to restart `HAP-python`. [#167](https://github.com/ikalchev/HAP-python/pull/167)
 
 ### Fixed
 - `AcessoryDriver.add_job` now correctly schedules coroutines wrapped in functools.partial.